To reduce complexity and improve readability and testability, move the form instance parsing code from XFormParser to a new class.
